Fix for degenerate polygons.
TopoJSON arcs always have at least two coordinates. If a polygon is constructed
from a single arc, and that arc only has two coordinates, then the resulting
reconstructed polygon previously only had a single coordinate. According to the
GeoJSON specification, this is invalid: "A LinearRing is closed LineString with
4 or more positions. The first and last positions are equivalent (they represent
equivalent points)." And so, now when degenerate polygons are constructed, the
first point is duplicated to the end to satisfy the specification.