Reverse the order of property transforms.
Previously, -p took the form source=target, mapping the input property named
"source" to the output property named "target". This was confusing because in
most programming languages, the equals operator (=) functions the other way,
copying the value of the right operand to the left operand. This commit changes
-p to the form target=source for consistency.

In addition, you can now coerce properties to numbers by prefixing the source
property name with a '+', similar to the --id-property. For example, -p bar=+FOO
will map the string property "FOO" to the number property "bar".