Join external properties to features.
Using the new -e (--external-properties) argument, you can now join a CSV or TSV
file of data to the topology as part of the conversion process. Each row in the
referenced DSV file is joined to the geometry objects in the generated topology
using the "id" attribute. This process happens after the generation of the
topology, and can thus the id of the geometry objects be configured using the
--id-property argument.

These external properties are transformed normally using the -p (--properties)
argument, and can thus be filtered and coerced to numbers as desired. Note that
by default, this means that any -e argument is ignored unless a -p argument is
also specified!