Remove topojson.object; bump to 1.0.0.
This wouldn’t have been backwards-compatible anyway because it changed the
behavior of topojson.object, removing propagation of id and properties. Rather
than drag the old method along, now seems as good a time as any to call this
version 1.0.0 of the client API reference—there are quite a few users already!