Scale and translate to fit a viewport.
Fixes #67. Using the --width, --height and --margin arguments, you can now scale
and translate TopoJSON in projected coordinates to fit a viewport. This makes it
easier to display TopoJSON without needing a projection!

This commit also adds an optional --bbox argument which you can use to include a
bbox property on the generated topology. The bbox has the same form as in
GeoJSON: [x0, y0, x1, y1];