Ensure cartesian areas are positive in pixels.
When we say "Cartesian" we really mean screen pixels, in which case the
vertical y-axis origin is at the top, not the bottom as in conventional
mathematical notation.  This affects the sign of the area calculation.

Forcing clockwise winding order in screen pixels requires us to change
the area sign to ensure that positive areas mean clockwise winding order
for both spherical and Cartesian coordinate systems.