Trivial topology generation (arcify).
Arcify is a “first-pass” topology construction that does not detect coincident
points and instead treats each input line as a discrete arc. This is a first
step towards constructing the topology. Each arc is represented as a continguous
slice of a shared point buffer; this will allow easy cutting of arcs later when
coincident points are detected.