Fix failing test.
The failing test contained two polygons, one with a CW and the other
with an equivalent but CCW ring, and each with a zero-area ring.  The
rings for both polygons are now automatically reordered so that the
zero-area rings are treated as holes.  The non-zero-area rings are
forced to be clockwise, since they are both exterior rings.  Finally,
the zero-area rings are removed by filter, resulting in both polygons
having the same clockwise ring.