Add --viewport-bbox.
Sometimes you want to specify the viewport bbox manually rather than relying on
the input. Although what I think I really want most commonly is to show the
entire world, and specifying that in coordinates is sort of a pain because some
projections have odd behavior for [-180,-90] and [180,90] (e.g., Mercator). So
it’d probably be better to compute the projected bounds of the sphere.