When stitching, only stitch exterior polygon rings.
There isn’t really a sensible situation in which interior polygon rings
would need stitching due to antimeridian cuts.  If a ring-shaped polygon
crosses the antimeridian, and it is cut across the hole, it would become
a multipolygon consisting of two single-ring polygons.

This also includes a fix for a duplicate endpoint that wasn’t being
dropped.