Simplify topojson.mesh; move topojson.merge.
Move merge to the TopoJSON client API. Rewrite mesh so that it is guaranteed to
preserve winding order of arcs (and output any leftover arcs that couldn’t be
stitched together).

Simplify the merge API: it is now just merge(geometries) where geometries is an
array of Polygon or MultiPolygon geometry objects. We could probably do
something like have merge(topology, geometries) return GeoJSON rather than
TopoJSON, for consistency with mesh(topology, object). I thought it was simpler
to have merge take an array of geometries, but I might also consider having it
take a GeometryCollection for consistency with mesh.

This commit also breaks the merge implementation, but I’m planning on rewriting
that next to preserve polygon semantics (hole-ring relationships).