Change pre-quantization defaults.
Pre-quantization is now disabled by default if only a single input is provided,
since the assumption is that when there’s only a single input, it is likely to
be consistent and thus we do not need fuzzy-matching to determine shared arcs.

Also, the -q command line option now only sets post-quantization, requiring you
to see pre-quantization explicitly if necessary.