Fix topojson-command-test on Linux.
It looks like Node’s child processes cannot always open /dev/stdout, at
least on Linux.  An error such as "UNKNOWN" or "ENXIO" is thrown.

This thread may be related:

  https://groups.google.com/d/topic/nodejs/SxNKLclbM5k/discussion

One workaround is to insert a pipe to cat, which allows the child
process to open /dev/stdout, and pipe to cat.  Then the output of cat
can be read by the test.

Fixes #204.