Remove 'else' before 'this.topojson = topojson'
This matches the current d3.js (as of 3.5.5) method of exporting the library.  I was running into trouble on GitHub's Atom-Shell (now called electron) when I tried to use topojson.  It would fail to import correctly and not create a topojson object.  The error message was:
```bash
[22095:0503/213544:INFO:CONSOLE(370)] "Uncaught ReferenceError: topojson is not defined", source: file:///homes/rmason/code/menubar/example/index.html (370)
```
and was included in index.html as follows:
```html
<script src="topojson.js"></script>
```